requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
apscheduler  # For scheduled updates
requests    # For YouTube API calls